8/3/21 UPDATE
Begin project. Worked through the first 3 chapters learning Python and Flask to create a microblog web app.

Created a Sign In form, and an Index page with a couple of hard coded posts to show formatting.

Next steps will be based upon the progression of the tutorial, and this repo will be updated to reflect changes.