Update 8/22/2021
This is the current state of the project after finishing Chapter 10 of the microblog tutorial. I am beginning to get ideas of how to change it up! For example, in a previous chapter I included pagination for user posts listed in the index. I would like to see what it would take for me to introduce an infinite scroll so posts won't have to be separated.